airflow/requirements.txt
scikit-learn==1.2.2
numpy==1.24.4
pandas
sqlalchemy
psycopg2-binary
python-dotenv
pyyaml
evidently==0.4.28